Changes to Key
Updated keyMatch() to encrypt given key and check if it matches the
actual key's encryption.
Removed enteredkey as an attribute since it doesn't need to be stored.
Key constructor now takes String key as a param and stores the encrypted
version of it.
Still need encryptKey() to actually encrypt the keys