Changes, TODOs, and deleted Box
I noticed inbox and outbox were the same thing with different names so the interface was unnecessary. Added inbox and outbox attributes to User. Added TODOs on most of the java files so we can see what else needs done. Other than that just general clean up of classes and filling in some of the methods that could be finished.